# InternPool: various optimizations